chore(refactor) Refactor my folders and status This chore re-arranges my folder structure and changes my status codes
