+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "test": "npm run delete && npm run migrate && set NODE_ENV=test && nyc --reporter=html --reporter=text mocha --require babel-core/register ./test --recursive --exit",
     "start": "npm run migrate && babel-node server.js --exec babel-node --presets es2015,stage-2",
-    "migrate": "node -r babel-register -r ./node_modules/dotenv/config models",
-    "delete": "node -r babel-register -r ./node_modules/dotenv/config test_models",
+    "migrate": "node -r babel-register -r ./node_modules/dotenv/config ./server/models",
+    "delete": "node -r babel-register -r ./node_modules/dotenv/config ./server/models/dropTables",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "author": "Obuladike Chisom",
